refactor backpack files
