+++ packages/wallets/backpack/src/window.ts
